requirements.txt
tushare>=1.2.89
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
loguru>=0.6.0 